+++ CHANGELOG.md
@@ -4,6 +4,18 @@ The changelog for `ReactiveLists`. Also see the [releases](https://github.com/pl
 
 ------
 
+0.1.2
+-----
+
+### Changed
+
+- Section and cell view models are now diffable by default. ([#119](https://github.com/plangrid/ReactiveLists/pull/119), [@jessesquires](https://github.com/jessesquires))
+Each provide default values for `diffingKey`, but you can customize them for your own needs or opt-out of automatic diffing.
+    - `CollectionSectionViewModel` protocol now inherits from `DiffableViewModel` protocol
+    - `CollectionCellViewModel` protocol now inherits from `DiffableViewModel` protocol
+    - ` TableSectionViewModel` protocol now inherits from `DiffableViewModel` protocol
+    - `TableCellViewModel` protocol now inherits from `DiffableViewModel` protocol
+
 0.1.1
 -----
 
@@ -22,7 +34,7 @@ This release closes the [0.1.1 milestone](https://github.com/plangrid/ReactiveLi
 
 ### New
 
-- You can now customize the cell insertion and deletion animations on `TableViewDriver`. (#115, @wickwirew)
+- You can now customize the cell insertion and deletion animations on `TableViewDriver`. ([#115](https://github.com/plangrid/ReactiveLists/pull/115), [@wickwirew](https://github.com/wickwirew))
 - `ViewRegistrationInfo` properties `reuseIdentifier` and `registrationMethod` are now public
 - `ViewRegistrationInfo` now conforms to `Equatable`
 - `SupplementaryViewInfo` now conforms to `Equatable`